Merging catalogs V2
Script for the second version of merging catalogs.
The objective is to make sure one event will be discarded by accident.
Part 1 2 and 3. Split the merged catalog that contains just double events (identified by two methods), and plot the waveforms and energy stack to verify if the number of events is 1 or 2.