Energy Stack detection Algorithm V5
The version 5 features a simplified algorithm (compared to the V4, published in my thesis) for the event detection using energy stack as its characteristic fuction.
Also, the quality control has been simplified to two tests:
QC1 verifies how many peaks in the CV above a certain threshold (ex: 0.3) and aproves potential events with three or less peaks.
QC2 verifies if the peaks are in a central position. This is necessary due an excessive number of false positives related to peaks at the slice borders.

The folders 'catalogs' and 'potential_event_plots' present a single example each.